pyproject.toml
[tool.poetry]
name = "ai-cron"
version = "0.2.0"
description = "Smart AI Cron Tool with Web UI (Ollama, Gemini, GPT Support)."
authors = ["Dev-D <dev@example.com>"]
readme = "README.md"
packages = [{include = "aicron"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
rich = "^13.0.0"
python-crontab = "^3.0.0"
litellm = "^1.0.0"
croniter = "^2.0.0"
ollama = "^0.1.0"
nicegui = "^1.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ai-cron = "aicron.main:app"
